refactor: share as_int id-coercion helper from apps.utils.fields
Move as_int out of apps.api.v2.utils into the shared apps.utils.fields so
non-API code can use it, and replace Node._parse_pk with it. as_int also
rejects booleans (_parse_pk wrongly coerced True to 1).
