+++ apps/utils/tests/test_fields.py
@@ -4,6 +4,18 @@ import pytest
 
 from apps.evaluations.models import EvaluationMessage, EvaluationResult, EvaluationRun, Evaluator
 from apps.utils.factories.evaluations import EvaluationConfigFactory
+from apps.utils.fields import as_int
+
+
+def test_as_int_coerces_and_drops_malformed():
+    assert as_int("7") == 7
+    assert as_int(7) == 7
+    assert as_int("abc") is None
+    assert as_int(None) is None
+    assert as_int("") is None
+    # booleans are not valid ids, even though ``int(True)`` would otherwise coerce to 1
+    assert as_int(True) is None
+    assert as_int(False) is None
 
 
 @pytest.mark.django_db()
